@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: protos/protos.proto
@@ -11,7 +12,6 @@
 //
 // Generated by ./generate_proto.sh
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -25,7 +25,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 }
 
 /// protos/barcode_format.proto
-enum BarcodeFormat: SwiftProtobuf.Enum {
+enum BarcodeFormat: SwiftProtobuf.Enum, Swift.CaseIterable {
   typealias RawValue = Int
   case unknown // = 0
   case aztec // = 1
@@ -81,13 +81,8 @@ enum BarcodeFormat: SwiftProtobuf.Enum {
     }
   }
 
-}
-
-#if swift(>=4.2)
-
-extension BarcodeFormat: CaseIterable {
   // The compiler won't synthesize support with the UNRECOGNIZED case.
-  static var allCases: [BarcodeFormat] = [
+  static let allCases: [BarcodeFormat] = [
     .unknown,
     .aztec,
     .code39,
@@ -101,12 +96,11 @@ extension BarcodeFormat: CaseIterable {
     .upce,
     .pdf417,
   ]
+
 }
 
-#endif  // swift(>=4.2)
-
 /// protos/scan_result.proto
-enum ResultType: SwiftProtobuf.Enum {
+enum ResultType: SwiftProtobuf.Enum, Swift.CaseIterable {
   typealias RawValue = Int
   case barcode // = 0
   case cancelled // = 1
@@ -135,23 +129,17 @@ enum ResultType: SwiftProtobuf.Enum {
     }
   }
 
-}
-
-#if swift(>=4.2)
-
-extension ResultType: CaseIterable {
   // The compiler won't synthesize support with the UNRECOGNIZED case.
-  static var allCases: [ResultType] = [
+  static let allCases: [ResultType] = [
     .barcode,
     .cancelled,
     .error,
   ]
+
 }
 
-#endif  // swift(>=4.2)
-
 /// protos/android_configuration.proto
-struct AndroidConfiguration {
+struct AndroidConfiguration: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -166,13 +154,16 @@ struct AndroidConfiguration {
   /// This parameter is only supported on Android devices.
   var useAutoFocus: Bool = false
 
+  /// This sets the string in the appBar (Android only)
+  var appBarTitle: String = String()
+
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 }
 
 /// protos/configuration.proto
-struct Configuration {
+struct Configuration: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -206,7 +197,7 @@ struct Configuration {
   fileprivate var _android: AndroidConfiguration? = nil
 }
 
-struct ScanResult {
+struct ScanResult: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -261,6 +252,7 @@ extension AndroidConfiguration: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
   static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
     1: .same(proto: "aspectTolerance"),
     2: .same(proto: "useAutoFocus"),
+    3: .same(proto: "appBarTitle"),
   ]
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -271,17 +263,21 @@ extension AndroidConfiguration: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
       switch fieldNumber {
       case 1: try { try decoder.decodeSingularDoubleField(value: &self.aspectTolerance) }()
       case 2: try { try decoder.decodeSingularBoolField(value: &self.useAutoFocus) }()
+      case 3: try { try decoder.decodeSingularStringField(value: &self.appBarTitle) }()
       default: break
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if self.aspectTolerance != 0 {
+    if self.aspectTolerance.bitPattern != 0 {
       try visitor.visitSingularDoubleField(value: self.aspectTolerance, fieldNumber: 1)
     }
     if self.useAutoFocus != false {
       try visitor.visitSingularBoolField(value: self.useAutoFocus, fieldNumber: 2)
+    }
+    if !self.appBarTitle.isEmpty {
+      try visitor.visitSingularStringField(value: self.appBarTitle, fieldNumber: 3)
     }
     try unknownFields.traverse(visitor: &visitor)
   }
@@ -289,6 +285,7 @@ extension AndroidConfiguration: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
   static func ==(lhs: AndroidConfiguration, rhs: AndroidConfiguration) -> Bool {
     if lhs.aspectTolerance != rhs.aspectTolerance {return false}
     if lhs.useAutoFocus != rhs.useAutoFocus {return false}
+    if lhs.appBarTitle != rhs.appBarTitle {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -321,6 +318,10 @@ extension Configuration: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementa
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    // The use of inline closures is to circumvent an issue where the compiler
+    // allocates stack space for every if/case branch local when no optimizations
+    // are enabled. https://github.com/apple/swift-protobuf/issues/1034 and
+    // https://github.com/apple/swift-protobuf/issues/1182
     if !self.strings.isEmpty {
       try visitor.visitMapField(fieldType: SwiftProtobuf._ProtobufMap<SwiftProtobuf.ProtobufString,SwiftProtobuf.ProtobufString>.self, value: self.strings, fieldNumber: 1)
     }
@@ -330,9 +331,9 @@ extension Configuration: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementa
     if self.useCamera != 0 {
       try visitor.visitSingularInt32Field(value: self.useCamera, fieldNumber: 3)
     }
-    if let v = self._android {
+    try { if let v = self._android {
       try visitor.visitSingularMessageField(value: v, fieldNumber: 4)
-    }
+    } }()
     if self.autoEnableFlash != false {
       try visitor.visitSingularBoolField(value: self.autoEnableFlash, fieldNumber: 5)
     }
